# backend/openapi.yaml
openapi: 3.0.3
info:
  title: FindBuyers Backend API
  version: 0.1.0
  description: |
    API contract for agent onboarding (user-details) and agency management.

    Auth is handled by Supabase; clients must send `Authorization: Bearer <supabase_jwt>`.
servers:
  - url: http://localhost:3000

tags:
  - name: Onboarding
  - name: Agencies

paths:
  /api/v1/user-details:
    post:
      tags: [Onboarding]
      summary: Create or update the authenticated user's user-details record
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertUserDetailsBody'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailsResponse'
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailsResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'

  /api/v1/agencies:
    post:
      tags: [Agencies]
      summary: Create an agency; the authenticated user becomes the first admin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgencyBody'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgencyResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/v1/agencies/{agencyUuid}:
    get:
      tags: [Agencies]
      summary: Get an agency
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/agencyUuid'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgencyResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Agencies]
      summary: Update an agency (admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/agencyUuid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgencyBody'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgencyResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Agencies]
      summary: Delete an agency (admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/agencyUuid'
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/agencies/{agencyUuid}/users:
    get:
      tags: [Agencies]
      summary: List agency users + roles
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/agencyUuid'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgencyUser'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Agencies]
      summary: Add an existing user to an agency (admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/agencyUuid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddAgencyUserBody'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    $ref: '#/components/schemas/AgencyUser'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/v1/agencies/{agencyUuid}/users/{userUuid}:
    patch:
      tags: [Agencies]
      summary: Update a user's role within an agency (admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/agencyUuid'
        - $ref: '#/components/parameters/userUuid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  $ref: '#/components/schemas/Role'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    $ref: '#/components/schemas/AgencyUser'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: "Conflict (e.g., would remove the last remaining admin)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Agencies]
      summary: Remove a user from an agency (admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/agencyUuid'
        - $ref: '#/components/parameters/userUuid'
      responses:
        '204':
          description: Removed
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: "Conflict (e.g., would remove the last remaining admin)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    agencyUuid:
      name: agencyUuid
      in: path
      required: true
      schema:
        type: string
        format: uuid
    userUuid:
      name: userUuid
      in: path
      required: true
      schema:
        type: string
        format: uuid

  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthenticated:
      description: Unauthenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          description: Optional validation or diagnostic details

    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          $ref: '#/components/schemas/Error'

    Role:
      type: string
      enum: [admin, member]

    AddAgencyUserBody:
      type: object
      required: [userUuid, role]
      properties:
        userUuid:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/Role'

    UpsertUserDetailsBody:
      type: object
      required: [fullName]
      properties:
        fullName:
          type: string
          minLength: 1
          maxLength: 200
        phone:
          type: string
          nullable: true

    UserDetails:
      type: object
      required: [id, uuid, authUserId, fullName, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int32
        uuid:
          type: string
          format: uuid
        authUserId:
          type: string
          format: uuid
        fullName:
          type: string
        phone:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    UserDetailsResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: '#/components/schemas/UserDetails'

    CreateAgencyBody:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        website:
          type: string
          format: uri
          nullable: true

    UpdateAgencyBody:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        website:
          type: string
          format: uri
          nullable: true

    Agency:
      type: object
      required: [id, uuid, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int32
        uuid:
          type: string
          format: uuid
        name:
          type: string
        website:
          type: string
          format: uri
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    AgencyResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: '#/components/schemas/Agency'

    AgencyUser:
      type: object
      required: [id, uuid, agencyUuid, userUuid, role, createdAt]
      properties:
        id:
          type: integer
          format: int32
        uuid:
          type: string
          format: uuid
        agencyUuid:
          type: string
          format: uuid
        userUuid:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/Role'
        createdAt:
          type: string
          format: date-time